Create Corn Production Analysis
Corn is the second most-produced crop in the world, and it’s more than just a staple in our diets. Corn is the primary U.S. feed grain, accounting for more than 95% of total feed grain production and use. As a result, animal feed makes up nearly 40% of the country’s corn usage. This is because corn is a rich source of carbohydrates, and in combination with protein from soybeans, it can make for an effective diet for livestock. From the sweetener in our coffees to the ethanol that powers our vehicles, corn has hundreds of uses.